# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/admin/calendars/{calendarId}:
    get:
      tags:
      - Admin - Calendars
      summary: Consulta el detalle de un calendario
      operationId: get_2
      parameters:
      - name: calendarId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CalendarDetailResponse'
    put:
      tags:
      - Admin - Calendars
      summary: 'Edita un calendario: sustituye reglas, festivos y jornadas especiales'
      operationId: update_2
      parameters:
      - name: calendarId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCalendarRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CalendarDetailResponse'
    delete:
      tags:
      - Admin - Calendars
      summary: Archiva un calendario (borrado logico)
      operationId: archive
      parameters:
      - name: calendarId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /api/v1/workdays/{workdayId}/corrections:
    post:
      tags:
      - Corrections
      operationId: request
      parameters:
      - name: workdayId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequestDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorrectionResponse'
...
  /api/v1/public/tenant-registrations:
    post:
      tags:
      - Public - Tenant registrations
      summary: Solicita el alta de una organización (RF-REG-001)
      operationId: request_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRegistrationRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TenantRegistrationAcceptedResponse'
  /api/v1/public/tenant-registrations/verify-email:
    post:
      tags:
      - Public - Tenant registrations
      summary: Confirma la dirección de correo de una solicitud (RF-REG-004)
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRegistrationEmailRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TenantRegistrationAcceptedResponse'
  /api/v1/public/tenant-registrations/resend-verification:
    post:
      tags:
      - Public - Tenant registrations
      summary: Reenvía el correo de verificación de una solicitud (RF-REG-004)
      operationId: resendVerification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TenantRegistrationAcceptedResponse'
  /api/v1/platform/tenants:
    get:
      tags:
      - Platform - Tenants
      summary: Lista los tenants, con filtro opcional por estado
      operationId: list
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedPlatformTenantsResponse'
    post:
      tags:
      - Platform - Tenants
      summary: Crea un tenant y su primer TENANT_ADMIN (solo PLATFORM_ADMIN)
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateTenantResponse'
  /api/v1/platform/tenants/{tenantId}/suspend:
    post:
      tags:
      - Platform - Tenants
      summary: Suspende un tenant activo con motivo obligatorio
      operationId: suspend
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuspendTenantRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlatformTenantDetailResponse'
  /api/v1/platform/tenants/{tenantId}/reactivate:
    post:
      tags:
      - Platform - Tenants
      summary: Reactiva un tenant suspendido
      operationId: reactivate
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlatformTenantDetailResponse'
  /api/v1/platform/tenants/{tenantId}/archive:
    post:
      tags:
      - Platform - Tenants
      summary: Archiva un tenant de forma permanente
      operationId: archive_1
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveTenantRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlatformTenantDetailResponse'
  /api/v1/platform/tenants/{tenantId}/activate:
    post:
      tags:
      - Platform - Tenants
      summary: Activa un tenant pendiente
      operationId: activate
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlatformTenantDetailResponse'
...
  /api/v1/app/absences:
    get:
      tags:
      - Absences
      operationId: list_2
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbsenceResponse'
    post:
      tags:
      - Absences
      operationId: request_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AbsenceResponse'
...
  /api/v1/employees/{employeeId}/activate:
    patch:
      tags:
      - Employees
      operationId: activate_1
      parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
...
  /api/v1/platform/tenants/{tenantId}:
    get:
      tags:
      - Platform - Tenants
      summary: Consulta el detalle de un tenant
      operationId: get_4
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PlatformTenantDetailResponse'
...
components:
  schemas:
...
    EmployeeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            type: string
          uniqueItems: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DayRulePayload:
      type: object
      description: 'Regla semanal: si el dia de la semana es laborable y cuantos minutos se esperan'
      properties:
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
          example: MONDAY
        working:
          type: boolean
        expectedMinutes:
          type: integer
          format: int32
          example: 480
          maximum: 1440
          minimum: 0
      required:
      - dayOfWeek
    HolidayPayload:
      type: object
      description: Festivo en una fecha local del calendario
      properties:
        date:
          type: string
          format: date
          example: '2026-01-06'
        name:
          type: string
          example: Reyes
          maxLength: 120
          minLength: 0
      required:
      - date
    SaveCalendarRequest:
      type: object
      description: Estado completo de un calendario laboral
      properties:
        name:
          type: string
          example: Calendario general
          maxLength: 120
          minLength: 0
        timezone:
          type: string
          description: Zona horaria IANA; por defecto la configurada
          example: Europe/Madrid
          maxLength: 60
          minLength: 0
        validFrom:
          type: string
          format: date
          example: '2026-01-01'
        validTo:
          type: string
          format: date
          description: Fin de vigencia inclusivo; null = indefinida
          example: '2026-12-31'
        dayRules:
          type: array
          items:
            $ref: '#/components/schemas/DayRulePayload'
        holidays:
          type: array
          items:
            $ref: '#/components/schemas/HolidayPayload'
        specialDays:
          type: array
          items:
            $ref: '#/components/schemas/SpecialDayPayload'
      required:
      - validFrom
    SpecialDayPayload:
      type: object
      description: Jornada especial que sustituye la regla semanal en una fecha concreta
      properties:
        date:
          type: string
          format: date
          example: '2026-12-24'
        name:
          type: string
          example: Jornada intensiva de Nochebuena
          maxLength: 120
          minLength: 0
        expectedMinutes:
          type: integer
          format: int32
          example: 300
          maximum: 1440
          minimum: 0
      required:
      - date
    CalendarDetailResponse:
      type: object
      description: Detalle de un calendario laboral
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        status:
          type: string
        dayRules:
          type: array
          items:
            $ref: '#/components/schemas/DayRulePayload'
        holidays:
          type: array
          items:
            $ref: '#/components/schemas/HolidayPayload'
        specialDays:
          type: array
          items:
            $ref: '#/components/schemas/SpecialDayPayload'
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CorrectionRequestDto:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 0
        proposedChanges:
          $ref: '#/components/schemas/ProposedChangesDto'
      required:
      - proposedChanges
    ProposedBreakDto:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
      required:
      - endedAt
      - startedAt
    ProposedChangesDto:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        breaks:
          type: array
          items:
            $ref: '#/components/schemas/ProposedBreakDto'
          maxItems: 24
          minItems: 0
      required:
      - breaks
      - endedAt
      - startedAt
    CorrectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workdayId:
          type: string
          format: uuid
        requestedBy:
          type: string
          format: uuid
        reason:
          type: string
        proposedChanges:
          $ref: '#/components/schemas/ProposedChangesResponse'
        status:
          type: string
        resolvedBy:
          type: string
          format: uuid
        resolvedAt:
          type: string
          format: date-time
        resolutionComment:
          type: string
        createdAt:
          type: string
          format: date-time
    ProposedBreakResponse:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    ProposedChangesResponse:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        breaks:
          type: array
          items:
            $ref: '#/components/schemas/ProposedBreakResponse'
...
    TenantRegistrationRequestBody:
      type: object
      properties:
        companyName:
          type: string
          maxLength: 200
          minLength: 0
        timezone:
          type: string
          minLength: 1
        firstName:
          type: string
          maxLength: 200
          minLength: 0
        lastName:
          type: string
          maxLength: 200
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 200
          minLength: 10
        acceptTerms:
          type: boolean
    TenantRegistrationAcceptedResponse:
      type: object
      properties:
        message:
          type: string
    VerifyRegistrationEmailRequest:
      type: object
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 0
    ResendVerificationRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 255
          minLength: 0
    CreateTenantRequest:
      type: object
      properties:
        tenantName:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
        adminEmail:
          type: string
          minLength: 1
        adminPassword:
          type: string
          maxLength: 2147483647
          minLength: 10
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
    CreateTenantResponse:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        adminUserId:
          type: string
          format: uuid
    SuspendTenantRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 0
    PlatformTenantDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
        suspendedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
        suspensionReason:
          type: string
    ArchiveTenantRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 0
...
    AbsenceRequestBody:
      type: object
      properties:
        absenceTypeId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
          maxLength: 500
          minLength: 0
      required:
      - absenceTypeId
      - endDate
      - startDate
    AbsenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        absenceTypeId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        status:
          type: string
        resolvedBy:
          type: string
          format: uuid
        resolvedAt:
          type: string
          format: date-time
        resolutionComment:
          type: string
        createdAt:
          type: string
          format: date-time
...
    PagedPlatformTenantsResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PlatformTenantSummaryResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PlatformTenantSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
        suspendedAt:
          type: string
          format: date-time
        userCount:
          type: integer
          format: int64
        lastAccessAt:
          type: string
          format: date-time
...